refactor iframe code into Preview component
